+++ src/components/task.js
@@ -1,5 +1,5 @@
 import {MONTH_NAMES} from "../const";
-import {formatTime} from "../utils";
+import {formatTime} from "../utils/common";
 import AbstractComponent from "./abstract-component";
 
 const createHashtagsMarkup = (hashtags) => {
