+++ src/utils/common.js
@@ -1,8 +1,3 @@
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
 export const getRandomArrayElement = (array) => {
   return array[Math.floor(Math.random() * array.length)];
 };
@@ -22,21 +17,4 @@ export const formatTime = (date) => {
   const interval = date.getHours() > 11 ? `pm` : `am`;
 
   return `${hours}:${minutes} ${interval}`;
-};
-
-export const createElement = (element) => {
-  const newElement = document.createElement(`div`);
-  newElement.insertAdjacentHTML(RenderPosition.BEFOREEND, element);
-  return newElement.firstChild;
-};
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
 };
