+++ src/utils/render.js
@@ -1,0 +1,21 @@
+export const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
+
+export const createElement = (element) => {
+  const newElement = document.createElement(`div`);
+  newElement.insertAdjacentHTML(RenderPosition.BEFOREEND, element);
+  return newElement.firstChild;
+};
+
+export const render = (container, element, place) => {
+  switch (place) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(element);
+      break;
+    case RenderPosition.BEFOREEND:
+      container.append(element);
+      break;
+  }
+};
